fix(slots): reviewer-blocking bugs in the visual floor plan (P2/P3)
Two fable-5 reviewers found blocking bugs; fixing all before merge.

CRITICAL — booth positions destroyed on the Frappe (production) path:
- exhibition_slot.json: pos_x/pos_y were Int; the visual map stores normalized
  0-1 floats, which Frappe cint-truncates to 0 → all booths stack in the corner.
  Now Float(precision 6). Dev (in-memory) masked it entirely.
- api/slots.py save_layout: int(pos.x) likewise truncated every admin save →
  float().

HIGH — 7 LB booths (LB5-1..LB6-2) were dead:
- They fail slots.ts SLOT_RE (one-letter grammar), so hold/status/layout routes
  400'd every action AND the layout editor couldn't re-save any layout with them.
- parseSlot stays strict (its tests require AA1→null); added isPlaceableSlotLabel
  for the broader placement grammar. hold/status routes now gate on layout[label]
  presence (the real source of truth); layout-save uses isPlaceableSlotLabel.

MEDIUM/LOW — seed.py: config now guards each field individually (no longer resets
an admin's currency/slots-per-letter when floor_image_url is empty); zone upsert
moved below the slot-count guard so a re-migrate can't resurrect a deleted zone.
FloorPlanMap: aspect ratio is a named constant, not a bare magic pair.

TS: typecheck clean, 44 slot tests green (incl. LB-label regression). Frappe:
Float/seed verified on a bench next.

+++ frappe_app/lebtech_partner_platform/lebtech_partner_platform/api/slots.py
@@ -105,8 +105,10 @@ def save_layout(**payload):
         else:
             doc = frappe.get_doc({"doctype": "Exhibition Slot", "slot_label": label, "status": "Available"})
         doc.zone = pos.get("zoneId")
-        doc.pos_x = int(pos.get("x") or 0)
-        doc.pos_y = int(pos.get("y") or 0)
+        # Normalized 0-1 image coordinates — float, NOT int. int() truncates every
+        # booth to 0 and stacks the whole map in the corner (both reviewers flagged).
+        doc.pos_x = float(pos.get("x") or 0)
+        doc.pos_y = float(pos.get("y") or 0)
         doc.price = float(prices.get(label) or 0)
         doc.is_active = 1 if label in active_set else 0
         doc.save() if not doc.is_new() else doc.insert()
